# Optimize performance (~4X faster build when images are up-to-date)