I've looked into Brick and wanted to figure out why building the tmrow repository was slower than expected. 

I got these timings locally when nothing changed and everything was already prepared and built:

    time brick -r prepare    : 48 seconds (same time if for example .npmrc was touched)
    time brick -r build      : 175 seconds

I would expect an almost instant prepare and build and time when the Docker images were up-to-date. 

It turns out that the overhead of running Buildit/Docker is rather slow for that many images – even though layers are cached. I tried to do some experiments with a plain Dockerfile and observed the same overhead.

I came up with two improvements some time ago... But now I managed to wrap it up into a PR. :) 

### UPDATE: Buildkit/Docker performance

The primary reasons for the performance issue with Buildkit on my machine are:
- general overhead in starting up Docker and running through each layer
- Slowest layers:  "resolve image config for" (0.4-1.2s), "[internal] load metadata for" (0.4s)



### Improvement 1: dependency change detection 

Idea: not running docker on `brick prepare` and `brick build` if input(s) and the BUILD.yaml file didn't change since the last image was tagged. I tried to think if this would work in all cases, and at least logically I cannot see why a build should be done if the inputs didn't change. Right? 🤔

UPDATE: originally I used the using file modified timestamps on input paths. But as file timestamps are not preserved in git, then CI wouldn't be able to cache anything as cloning the repository would update the modified timestamp. Instead we now hash all dependencies (inputs and BUILD.yaml file). The hashes are stored in a `~/.brick-cache.json` file.

When everything is prepared/built already:

    time brick -r prepare    : 1.5 seconds
    time brick -r build      : 66 seconds

So when everything is up-to-date: brick build time is 2.5X faster and prepare time is 32X faster. 🚀


### Improvement 2: output extraction using get_archive

The output extraction after a build is very slow. I noticed a comment around this to use `get_archive`. Doing so improves the performance a bit:

    time brick -r build      : 144 seconds (without improvement 1)


### Improvement 1 + 2:

When everything is already built:

    time brick -r build      : 46 seconds

So we end up with: brick build time is 4X faster and prepare time is still 24X faster. 🚀

But the build time should be instant when nothing changed, right? It turns out that collecting outputs is still very slow and is always done. 

When disabling any collecting of outputs:
    time brick -r build      : 5 seconds


### Conclusion

Of course these timing are made when everything is up-to-date. But as most PRs only touches a few images, then the performance improvement should be substantial on CI.

Next step could be to be more clever when doing output extraction using hashing. 🌤 But maybe that performance increase isn't worth the added complexities.


### TODO

- [x] Sanity check my assumption that prepare and build should not be done unless inputs or BUILD.yaml file changed
- [x] Performance test on CI (modified timestamp)
- [ ] Figure out how the `~/.brick-cache.json` can be set up on CI + how it works between branches... Maybe we need to store the digest? Is the local docker registry shared between builds?
- [ ] Performance test on CI (new hashing version + preserving the `~/.brick-cache.json` file)
- [ ] Solution for stale cache file when images are being pruned by Drone cronjob or similar... Probably we need to store the digest in the cache
- [ ] Verify that we didn't break anything (mostly this would be stale builds, i.e. not building when we should)