@madsnedergaard 2020-08-24
This is epic, I've spent so much time trying to figure out why it's running so slowly in CI (but fine locally) 😭 

Will get straight to testing it now!