@skovhus 2020-08-24
> This is epic, I've spent so much time trying to figure out why it's running so slowly in CI (but fine locally) 😭

Glad to hear! All timings here are local. I haven't tested on CI yet (and don't plan to do that before September).

So any help testing would be greatness! 👏