@corradio 2020-08-25
This is quite a clever way to optimise the build. I'll let Mads review it but from my point of view it looks good to me!