@corradio 2020-08-25
I'm still surprised that buildkit is not doing this by itself. Maybe worth opening up an issue on https://github.com/moby/buildkit/ or reaching out to the folks.