@skovhus 2020-08-26
> I'm still surprised that buildkit is not doing this by itself. 

Yes, we should debug this a bit more. I observed a diffing overhead on a plain Dockerfile, but we could investigate further.