@skovhus 2020-08-27
Replaced by https://github.com/tmrowco/brick/pull/45 and https://github.com/tmrowco/brick/pull/46